requirements.txt
# Core dependencies
google-genai
feedparser
requests
beautifulsoup4
python-dotenv
duckduckgo-search
